requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langchain-text-splitters
sentence-transformers
openai
faiss-cpu
scikit-learn
gensim
pandas
numpy
PyMuPDF
pdfplumber
pytesseract
Pillow
camelot-py
tqdm
huggingface_hub[hf_xet]>=0.23
transformers>=4.40.0
accelerate>=0.30.0
torch>=2.0.0

